Moved things around to deal with some pytest stuff I don't understand
